refactor(Entity): Move Entity instance methods to entity.js
