+++ lib/entity.js
@@ -8,9 +8,9 @@ const arrify = require('arrify');
 const datastoreSerializer = require('./serializer').Datastore;
 const defaultValues = require('./helpers/defaultValues');
 const { errorCodes } = require('./errors');
-const { populateHelpers } = require('./helpers');
+const { validation, populateHelpers } = require('./helpers');
 
-const { addPathToPopulateRefs, populateFactory } = populateHelpers;
+const { populateFactory } = populateHelpers;
 
 class Entity {
     constructor(data, id, ancestors, namespace, key) {
@@ -43,6 +43,176 @@ class Entity {
 
         // add middleware defined on Schena
         registerHooksFromSchema(this);
+    }
+
+    save(transaction, opts = {}) {
+        this.__hooksEnabled = true;
+        const _this = this;
+        const options = {
+            method: 'upsert',
+            ...opts,
+        };
+
+        let { error } = validateEntityData();
+
+        if (!error) {
+            ({ error } = validateMethod(options.method));
+        }
+
+        if (error) {
+            return Promise.reject(error);
+        }
+
+        this.entityData = prepareData.call(this);
+
+        const entity = datastoreSerializer.toDatastore(this);
+        entity.method = options.method;
+
+        if (!transaction) {
+            return this.gstore.ds.save(entity).then(onSuccess);
+        }
+
+        if (transaction.constructor.name !== 'Transaction') {
+            return Promise.reject(new Error('Transaction needs to be a gcloud Transaction'));
+        }
+
+        addPostHooksTransaction.call(this);
+        transaction.save(entity);
+
+        return Promise.resolve(this);
+
+        // --------------------------
+
+        function onSuccess() {
+            /**
+             * Make sure to clear the cache for this Entity Kind
+             */
+            if (_this.constructor.__hasCache(options)) {
+                return _this.constructor.clearCache()
+                    .then(() => _this)
+                    .catch((err) => {
+                        let msg = 'Error while clearing the cache after saving the entity.';
+                        msg += 'The entity has been saved successfully though. ';
+                        msg += 'Both the cache error and the entity saved have been attached.';
+                        const cacheError = new Error(msg);
+                        cacheError.__entity = _this;
+                        cacheError.__cacheError = err;
+                        throw cacheError;
+                    });
+            }
+
+            return _this;
+        }
+
+        function validateEntityData() {
+            if (_this.schema.options.validateBeforeSave) {
+                return _this.validate();
+            }
+
+            return {};
+        }
+
+        function validateMethod(method) {
+            const allowed = {
+                update: true,
+                insert: true,
+                upsert: true,
+            };
+
+            return !allowed[method]
+                ? { error: new Error('Method must be either "update", "insert" or "upsert"') }
+                : { error: null };
+        }
+
+        /**
+         * Process some basic formatting to the entity data before save
+         * - automatically set the modifiedOn property to current date (if exists on schema)
+         * - convert object with latitude/longitude to Datastore GeoPoint
+         */
+        function prepareData() {
+            updateModifiedOn.call(this);
+            convertGeoPoints.call(this);
+
+            return this.entityData;
+
+            //--------------------------
+
+            /**
+             * If the schema has a "modifiedOn" property we automatically
+             * update its value to the current dateTime
+            */
+            function updateModifiedOn() {
+                if ({}.hasOwnProperty.call(this.schema.paths, 'modifiedOn')) {
+                    this.entityData.modifiedOn = new Date();
+                }
+            }
+
+            /**
+             * If the entityData has some property of type 'geoPoint'
+             * and its value is an js object with "latitude" and "longitude"
+             * we convert it to a datastore GeoPoint.
+            */
+            function convertGeoPoints() {
+                if (!{}.hasOwnProperty.call(this.schema.__meta, 'geoPointsProps')) {
+                    return;
+                }
+
+                this.schema.__meta.geoPointsProps.forEach((property) => {
+                    if ({}.hasOwnProperty.call(_this.entityData, property)
+                        && _this.entityData[property] !== null
+                        && _this.entityData[property].constructor.name !== 'GeoPoint') {
+                        _this.entityData[property] = _this.gstore.ds.geoPoint(_this.entityData[property]);
+                    }
+                });
+            }
+        }
+
+        /**
+         * If it is a transaction, we create a hooks.post array that will be executed
+         * when transaction succeeds by calling transaction.execPostHooks() (returns a Promises)
+         */
+        function addPostHooksTransaction() {
+            // disable (post) hooks, we will only trigger them on transaction succceed
+            this.__hooksEnabled = false;
+            this.constructor.hooksTransaction.call(
+                _this,
+                transaction,
+                this.__posts
+                    ? this.__posts.save
+                    : undefined
+            );
+        }
+    }
+
+    validate() {
+        const { schema, entityKind } = this;
+        const { entityData } = this;
+        const key = this.entityData[this.gstore.ds.KEY]; // save the Key
+
+        /**
+         * If not a Joi schema, we sanitize the entityData.
+         * If it's a Joi, it will be done automatically when validating.
+         */
+        if (is.undef(schema._joi)) {
+            this.entityData = this.constructor.sanitize(entityData, { disabled: ['write'] });
+        }
+
+        const validationResult = validation.validate(
+            this.entityData,
+            schema,
+            entityKind,
+            this.gstore.ds
+        );
+
+        /**
+         * If it's a Joi schema, make sure to update the entityData object
+         */
+        if (is.defined(schema._joi)) {
+            this.entityData = validationResult.value;
+        }
+
+        this.entityData[this.gstore.ds.KEY] = key; // put the Key back
+        return validationResult;
     }
 
     plain(options) {
@@ -123,16 +293,14 @@ class Entity {
         }
     }
 
-    // TODO: Move all entity method HERE
     populate(path, propsToSelect) {
-        const paths = path ? arrify(path) : this.constructor.getEntitiesRefsFromSchema();
         const refsToPopulate = [];
-        paths.forEach(p => addPathToPopulateRefs(p, propsToSelect, refsToPopulate));
 
-        const promise = Promise.resolve()
-            .then(() => this.constructor.populate(refsToPopulate)(this));
+        const promise = Promise.resolve(this)
+            .then(this.constructor.populate(refsToPopulate));
 
         promise.populate = populateFactory(refsToPopulate, promise, this.constructor);
+        promise.populate(path, propsToSelect);
         return promise;
     }
