+++ lib/helpers/populateHelpers.js
@@ -16,11 +16,11 @@ const arrify = require('arrify');
  * }
  *
  * To fetch the "address", the "user" and the user's "conmpany", the array of refs
- * to retrieve will be
+ * to retrieve will have the following shape
  *
  * [
- *  [{ path: 'user', select: ['*'] }, [ path: 'address', select: ['*'] ], // tree depth = 0
- *  [{ path: 'user.company', select: ['*'] }], // tree depth = 1
+ *  [{ path: 'user', select: ['*'] }, [ path: 'address', select: ['*'] ], // tree depth at level 0
+ *  [{ path: 'user.company', select: ['*'] }], // tree depth at level 1 (will be fetched after level 0 has been fetched)
  * ]
  */
 const addPathToPopulateRefs = (initialPath, _select = ['*'], refs) => {
