+++ lib/model.js
@@ -10,53 +10,11 @@ const set = require('lodash.set');
 
 const Entity = require('./entity');
 const Query = require('./query');
-const datastoreSerializer = require('./serializer').Datastore;
 const { GstoreError, errorCodes } = require('./errors');
-const { validation } = require('./helpers');
 const { populateHelpers } = require('./helpers');
 
 const { keyToString } = dsAdapter;
 const { populateFactory } = populateHelpers;
-
-const sanitize = (data, schema, options = { disabled: [] }) => {
-    if (!is.object(data)) {
-        return null;
-    }
-
-    const isJoiSchema = !is.undef(schema._joi);
-
-    let sanitized;
-    let joiOptions;
-    if (isJoiSchema) {
-        sanitized = schema._joi.validate(data).value;
-        joiOptions = schema.options.joi.options || {};
-    } else {
-        sanitized = Object.assign({}, data);
-    }
-
-    const isSchemaExplicitOnly = isJoiSchema
-        ? !joiOptions.allowUnknown
-        : schema.options.explicitOnly !== false;
-
-    const isWriteDisabled = options.disabled.includes('write');
-    let schemaHasProperty;
-    let isPropWritable;
-
-    Object.keys(data).forEach((k) => {
-        schemaHasProperty = {}.hasOwnProperty.call(schema.paths, k);
-        isPropWritable = schemaHasProperty
-            ? schema.paths[k].write !== false
-            : true;
-
-        if ((isSchemaExplicitOnly && !schemaHasProperty) || (!isPropWritable && !isWriteDisabled)) {
-            delete sanitized[k];
-        } else if (sanitized[k] === 'null') {
-            sanitized[k] = null;
-        }
-    });
-
-    return sanitized;
-};
 
 class Model extends Entity {
     static compile(kind, schema, gstore) {
@@ -660,8 +618,46 @@ class Model extends Entity {
      * Sanitize user data before saving to Datastore
      * @param data : userData
      */
-    static sanitize(data) {
-        return sanitize(data, this.schema);
+    static sanitize(data, options = { disabled: [] }) {
+        const { schema } = this;
+
+        if (!is.object(data)) {
+            return null;
+        }
+
+        const isJoiSchema = !is.undef(schema._joi);
+
+        let sanitized;
+        let joiOptions;
+        if (isJoiSchema) {
+            sanitized = schema._joi.validate(data).value;
+            joiOptions = schema.options.joi.options || {};
+        } else {
+            sanitized = Object.assign({}, data);
+        }
+
+        const isSchemaExplicitOnly = isJoiSchema
+            ? !joiOptions.allowUnknown
+            : schema.options.explicitOnly !== false;
+
+        const isWriteDisabled = options.disabled.includes('write');
+        let schemaHasProperty;
+        let isPropWritable;
+
+        Object.keys(data).forEach((k) => {
+            schemaHasProperty = {}.hasOwnProperty.call(schema.paths, k);
+            isPropWritable = schemaHasProperty
+                ? schema.paths[k].write !== false
+                : true;
+
+            if ((isSchemaExplicitOnly && !schemaHasProperty) || (!isPropWritable && !isWriteDisabled)) {
+                delete sanitized[k];
+            } else if (sanitized[k] === 'null') {
+                sanitized[k] = null;
+            }
+        });
+
+        return sanitized;
     }
 
     /**
@@ -817,176 +813,6 @@ class Model extends Entity {
         return Object.entries(this.schema.paths)
             .filter(({ 1: pathConfig }) => pathConfig.type === 'entityKey')
             .map(({ 0: ref }) => ref);
-    }
-
-    save(transaction, _options = {}) {
-        this.__hooksEnabled = true;
-        const _this = this;
-        const options = {
-            method: 'upsert',
-            ..._options,
-        };
-
-        let { error } = validateEntityData();
-
-        if (!error) {
-            ({ error } = validateMethod(options.method));
-        }
-
-        if (error) {
-            return Promise.reject(error);
-        }
-
-        this.entityData = prepareData.call(this);
-
-        const entity = datastoreSerializer.toDatastore(this);
-        entity.method = options.method;
-
-        if (!transaction) {
-            return this.gstore.ds.save(entity).then(onSuccess);
-        }
-
-        if (transaction.constructor.name !== 'Transaction') {
-            return Promise.reject(new Error('Transaction needs to be a gcloud Transaction'));
-        }
-
-        addPostHooksTransaction.call(this);
-        transaction.save(entity);
-
-        return Promise.resolve(this);
-
-        // --------------------------
-
-        function onSuccess() {
-            /**
-             * Make sure to clear the cache for this Entity Kind
-             */
-            if (_this.constructor.__hasCache(options)) {
-                return _this.constructor.clearCache()
-                    .then(() => _this)
-                    .catch((err) => {
-                        let msg = 'Error while clearing the cache after saving the entity.';
-                        msg += 'The entity has been saved successfully though. ';
-                        msg += 'Both the cache error and the entity saved have been attached.';
-                        const cacheError = new Error(msg);
-                        cacheError.__entity = _this;
-                        cacheError.__cacheError = err;
-                        throw cacheError;
-                    });
-            }
-
-            return _this;
-        }
-
-        function validateEntityData() {
-            if (_this.schema.options.validateBeforeSave) {
-                return _this.validate();
-            }
-
-            return {};
-        }
-
-        function validateMethod(method) {
-            const allowed = {
-                update: true,
-                insert: true,
-                upsert: true,
-            };
-
-            return !allowed[method]
-                ? { error: new Error('Method must be either "update", "insert" or "upsert"') }
-                : { error: null };
-        }
-
-        /**
-         * Process some basic formatting to the entity data before save
-         * - automatically set the modifiedOn property to current date (if exists on schema)
-         * - convert object with latitude/longitude to Datastore GeoPoint
-         */
-        function prepareData() {
-            updateModifiedOn.call(this);
-            convertGeoPoints.call(this);
-
-            return this.entityData;
-
-            //--------------------------
-
-            /**
-             * If the schema has a "modifiedOn" property we automatically
-             * update its value to the current dateTime
-            */
-            function updateModifiedOn() {
-                if ({}.hasOwnProperty.call(this.schema.paths, 'modifiedOn')) {
-                    this.entityData.modifiedOn = new Date();
-                }
-            }
-
-            /**
-             * If the entityData has some property of type 'geoPoint'
-             * and its value is an js object with "latitude" and "longitude"
-             * we convert it to a datastore GeoPoint.
-            */
-            function convertGeoPoints() {
-                if (!{}.hasOwnProperty.call(this.schema.__meta, 'geoPointsProps')) {
-                    return;
-                }
-
-                this.schema.__meta.geoPointsProps.forEach((property) => {
-                    if ({}.hasOwnProperty.call(_this.entityData, property)
-                        && _this.entityData[property] !== null
-                        && _this.entityData[property].constructor.name !== 'GeoPoint') {
-                        _this.entityData[property] = _this.gstore.ds.geoPoint(_this.entityData[property]);
-                    }
-                });
-            }
-        }
-
-        /**
-         * If it is a transaction, we create a hooks.post array that will be executed
-         * when transaction succeeds by calling transaction.execPostHooks() (returns a Promises)
-         */
-        function addPostHooksTransaction() {
-            // disable (post) hooks, we will only trigger them on transaction succceed
-            this.__hooksEnabled = false;
-            this.constructor.hooksTransaction.call(
-                _this,
-                transaction,
-                this.__posts
-                    ? this.__posts.save
-                    : undefined
-            );
-        }
-    }
-
-    validate() {
-        const { schema, entityKind } = this;
-        const { entityData } = this;
-        const key = this.entityData[this.gstore.ds.KEY]; // save the Key
-
-        /**
-         * If not a Joi schema, we sanitize the entityData.
-         * If it's a Joi, it will be done automatically when validating.
-         */
-        if (is.undef(schema._joi)) {
-            this.entityData = sanitize(entityData, schema, { disabled: ['write'] });
-        }
-
-        const validationResult = validation.validate(
-            this.entityData,
-            schema,
-            entityKind,
-            this.gstore.ds
-        );
-
-        /**
-         * If it's a Joi schema, make sure to update the entityData object
-         */
-        if (is.defined(schema._joi)) {
-            this.entityData = validationResult.value;
-        }
-
-        this.entityData[this.gstore.ds.KEY] = key; // put the Key back
-        return validationResult;
     }
 
     // ------------------------------------------------------------------------
